# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ReservationStatus = typeof ReservationStatus[keyof typeof ReservationStatus];


export const ReservationStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface Reservation {
  id: string;
  name: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  partySize: string;
  specialRequests?: string;
  status: ReservationStatus;
  createdAt: string;
}

export interface CreateReservationInput {
  name: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  partySize: string;
  specialRequests?: string;
}

export type UpdateReservationInputStatus = typeof UpdateReservationInputStatus[keyof typeof UpdateReservationInputStatus];


export const UpdateReservationInputStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface UpdateReservationInput {
  status?: UpdateReservationInputStatus;
  name?: string;
  email?: string;
  phone?: string;
  date?: string;
  time?: string;
  partySize?: string;
  specialRequests?: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminLoginResponse {
  accessToken: string;
}