# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reservations
    description: Reservation operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reservations:
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/reservations:
    get:
      operationId: listReservations
      tags: [admin]
      summary: List all reservations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/reservations/{id}:
    patch:
      operationId: updateReservation
      tags: [admin]
      summary: Update a reservation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationInput"
      responses:
        "200":
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteReservation
      tags: [admin]
      summary: Delete a reservation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Reservation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        date:
          type: string
        time:
          type: string
        partySize:
          type: string
        specialRequests:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - date
        - time
        - partySize
        - status
        - createdAt

    CreateReservationInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        date:
          type: string
        time:
          type: string
        partySize:
          type: string
        specialRequests:
          type: string
      required:
        - name
        - email
        - phone
        - date
        - time
        - partySize

    UpdateReservationInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        date:
          type: string
        time:
          type: string
        partySize:
          type: string
        specialRequests:
          type: string

    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AdminLoginResponse:
      type: object
      properties:
        accessToken:
          type: string
      required:
        - accessToken