# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Create a reservation
 */
export const CreateReservationBody = zod.object({
  "name": zod.string(),
  "email": zod.string(),
  "phone": zod.string(),
  "date": zod.string(),
  "time": zod.string(),
  "partySize": zod.string(),
  "specialRequests": zod.string().optional()
})


/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  "email": zod.string(),
  "password": zod.string()
})

export const AdminLoginResponse = zod.object({
  "accessToken": zod.string()
})


/**
 * @summary List all reservations
 */
export const ListReservationsResponseItem = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "email": zod.string(),
  "phone": zod.string(),
  "date": zod.string(),
  "time": zod.string(),
  "partySize": zod.string(),
  "specialRequests": zod.string().optional(),
  "status": zod.enum(['pending', 'confirmed', 'cancelled', 'completed']),
  "createdAt": zod.string()
})
export const ListReservationsResponse = zod.array(ListReservationsResponseItem)


/**
 * @summary Update a reservation
 */
export const UpdateReservationParams = zod.object({
  "id": zod.coerce.string()
})

export const UpdateReservationBody = zod.object({
  "status": zod.enum(['pending', 'confirmed', 'cancelled', 'completed']).optional(),
  "name": zod.string().optional(),
  "email": zod.string().optional(),
  "phone": zod.string().optional(),
  "date": zod.string().optional(),
  "time": zod.string().optional(),
  "partySize": zod.string().optional(),
  "specialRequests": zod.string().optional()
})

export const UpdateReservationResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "email": zod.string(),
  "phone": zod.string(),
  "date": zod.string(),
  "time": zod.string(),
  "partySize": zod.string(),
  "specialRequests": zod.string().optional(),
  "status": zod.enum(['pending', 'confirmed', 'cancelled', 'completed']),
  "createdAt": zod.string()
})


/**
 * @summary Delete a reservation
 */
export const DeleteReservationParams = zod.object({
  "id": zod.coerce.string()
})